fix x64 compile

+++ formats/png.h
@@ -32,12 +32,12 @@ class Png : Format
 
 public:
 
-    Png(void *p, uint32_t s = 0) : Format(p, s) {}
+    Png(void *p, size_t s = 0) : Format(p, s) {}
     // using Format::Format;
     // VS2013 does not support C++11 inheriting constructors
 
 
-    uint32_t Leanify(uint32_t size_leanified = 0);
+    size_t Leanify(size_t size_leanified = 0);
 
     static const unsigned char header_magic[8];
